feat(toggle-change-month) Add changeMonthOnDayClick prop to DPI
This commit adds a new prop, changeMonthOnDayClick, to DayPickerInput. The prop allows
the user to specify if the DayPickerInput should update the `month` prop on
the DayPicker it renders when a day is clicked. It is defaulted to true.

This allows the user to disable the default behavior of a DayPicker that is
displaying multiple months from jumping forward to the month of the day that was clicked.