Only announce month when the month changes
aria-live as it is implemented currently causes usability issues for this component.

You cannot render it on a page without it interfering what what the screen reader user will hear. `aria-live` forces the screen reader to be high-jacked by the content on this element regardless of the content order. This is a bad user experience. Instead, the sreen-reader should wait until the screen reader cursor reached the datepicker to announce this content. Additionally, announce it when the user interacts with the next and previous butons.

This causes accessibility issues for a number of use-cases where we defer the mounting of the datepicker.

1. Datepicker is in a modal
2. Datepicker is in a dropdown as part of a combobox
3. Virtualization -- when unmounting components that are out of the view to reduce DOM overhead.